refactor: split the analysis result contracts by workflow
`analysisResults.ts` held two unrelated workflows' shapes, where the rest
of the package is one module per domain. NuVs and pathoscope get their own,
and `NuvsBlast` moves out of `analyses.ts` to sit with the contigs it is
requested against.

`Coordinate` goes with pathoscope: both `align` fields, both coverage
charts and the Visvalingam-Whyatt simplifier are its only consumers, so a
general module for it would have one member and no second caller.

The package has a single entry point and nothing imports a subpath, so no
call site changes.
